Move PerformativeException into a suitable package
