+++ src/main/java/performative/exception/PerformativeException.java
@@ -1,4 +1,4 @@
-package performative;
+package performative.exception;
 
 public class PerformativeException extends Exception {
     public PerformativeException(String message) {
